requirements.txt
python-telegram-bot[job-queue]==21.7
python-dotenv==1.0.1
PyYAML==6.0.2
aiosqlite==0.20.0
